Context was implemented
Use reducer to communicate with state
