+++ App.js
@@ -1,21 +1,9 @@
-import { Alert, StatusBar, StyleSheet, View, Platform } from "react-native";
 import React, { useState } from "react";
-import Constants from "expo-constants";
 import * as Font from "expo-font";
 import { AppLoading } from "expo";
 
-import NavBar from "./src/components/NavBar";
-import MainScreen from "./src/screens/MainScreen";
-import TodoScreen from "./src/screens/TodoScreen";
-import { THEME } from "./src/theme";
-
-function MyStatusBar({ backgroundColor, ...props }) {
-  return (
-    <View style={{ backgroundColor, height: Constants.statusBarHeight }}>
-      <StatusBar translucent backgroundColor={backgroundColor} {...props} />
-    </View>
-  );
-}
+import MainLayout from "./src/MainLayout";
+import TodoState from "./src/context/todo/TodoState";
 
 async function loadApp() {
   await Font.loadAsync({
@@ -26,19 +14,7 @@ async function loadApp() {
 
 export default function App() {
   const [isReady, setIsReady] = useState(false);
-  const [todoId, setTodoId] = useState(null);
-  const [todos, setTodos] = useState([
-    { id: 1, title: "test 1" },
-    { id: 2, title: "test 2" },
-    { id: 3, title: "test 3" },
-    { id: 4, title: "test 4" },
-    { id: 5, title: "test 5" },
-    { id: 6, title: "test 6" },
-    { id: 7, title: "test 7" },
-    { id: 8, title: "test 8" },
-    { id: 9, title: "test 9" },
-  ]);
-
+  
   if (!isReady) {
     return (
       <AppLoading
@@ -49,86 +25,9 @@ export default function App() {
     );
   }
 
-  const addTodo = title => {
-    setTodos(prev => [
-      {
-        id: Date.now().toString(),
-        title,
-      },
-      ...prev,
-    ]);
-  };
-
-  const removeTodo = id => {
-    Alert.alert(
-      "Todo delete",
-      "Are you sure?",
-      [
-        {
-          text: "Cancel",
-          style: "cancel",
-        },
-        {
-          text: "Delete",
-          style: "destructive",
-          onPress: () => {
-            setTodoId(null);
-            setTodos(prev => prev.filter(todo => todo.id !== id));
-          },
-        },
-      ],
-      { cancelable: false },
-    );
-  };
-
-  const updateTodo = (id, title) => {
-    setTodos(prev =>
-      prev.map(todo => {
-        if (todo.id === id) {
-          todo.title = title;
-        }
-        return todo;
-      }),
-    );
-  };
-
-  let content = (
-    <MainScreen
-      addTodo={addTodo}
-      removeTodo={removeTodo}
-      todos={todos}
-      openTodo={setTodoId}
-    />
-  );
-
-  if (todoId) {
-    const selectedTodo = todos.find(todo => todo.id === todoId);
-    content = (
-      <TodoScreen
-        goBack={() => setTodoId(null)}
-        todo={selectedTodo}
-        onRemove={removeTodo}
-        onSave={updateTodo}
-      />
-    );
-  }
-
   return (
-    <View style={{ flex: 1 }}>
-      <MyStatusBar
-        backgroundColor={Platform.OS === "ios" ? "#fff" : THEME.MAIN_COLOR}
-        barStyle={Platform.OS === "ios" ? "dark-conten" : "light-content"}
-      />
-      <NavBar title='Todo App' />
-      <View style={styles.container}>{content}</View>
-    </View>
+    <TodoState>
+      <MainLayout />
+    </TodoState>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    paddingHorizontal: THEME.PADDING_HORIZONTAL,
-    paddingVertical: 10,
-    flex: 1,
-  },
-});
