+++ README.md
@@ -1,0 +1,22 @@
+## This is my first React Native project where I develope a simple TODO app with various features
+
+##
+### Installation
+Clone this [repository](https://github.com/fufylev/react_native_todo)
+
+
+First you must install [expo](https://expo.io/)
+```
+npm install expo-cli --global
+```
+
+Second
+```
+cd react_native_todo
+npm install
+expo start
+```
+
+You must have an "expo" application on your mobile device in order to see how this app works
+
+
