+++ src/context/todo/todoContext.js
@@ -1,0 +1,4 @@
+import {createContext} from 'react';
+
+export const TodoContext = createContext();
+
