+++ src/context/types.js
@@ -1,0 +1,3 @@
+export const ADD_TODO = 'ADD_TODO';
+export const UPDATE_TODO = 'UPDATE_TODO';
+export const REMOVE_TODO = 'REMOVE_TODO';
